+++ src/components/event/Daily.vue
@@ -2,20 +2,6 @@
   <div class="event-daily">
     <div class="event-daily__section">
       <div class="event-daily__label">
-        {{ $t('event.repetitions.since') }}
-      </div>
-      <div class="event-daily__since">
-        {{ originalEvent.payload.timestamp | prettyFullDate }}
-        <span
-          v-if="daysRepeating > 1"
-          class="event-daily__since-days"
-        >
-          — {{ $tc('event.repetitions.days', daysRepeating) }}
-        </span>
-      </div>
-    </div>
-    <div class="event-daily__section">
-      <div class="event-daily__label">
         {{ $t('event.daily.lastTwoWeeks') }}
       </div>
       <Chart :points="chartData" />
@@ -27,7 +13,7 @@
 import Vue from 'vue';
 import Chart from '../events/Chart.vue';
 import { GET_CHART_DATA } from '../../store/modules/events/actionTypes';
-import { HawkEvent, HawkEventRepetition } from '../../types/events';
+import { HawkEvent } from '../../types/events';
 import { EventChartItem } from '../../types/chart';
 
 export default Vue.extend({
@@ -59,30 +45,6 @@ export default Vue.extend({
        */
       chartData: [],
     };
-  },
-  computed: {
-    /**
-     * Get the the first event entity
-     */
-    originalEvent(): HawkEventRepetition {
-      return this.event.repetition;
-    },
-
-    /**
-     * Return concrete date
-     */
-    daysRepeating(): number {
-      if (!this.originalEvent) {
-        return 0;
-      }
-
-      const now = (new Date()).getTime();
-      const eventTimestamp = this.originalEvent.payload.timestamp * 1000;
-      const firstOccurrence = (new Date(eventTimestamp).getTime());
-      const differenceInDays = (now - firstOccurrence) / (1000 * 3600 * 24);
-
-      return Math.round(differenceInDays);
-    },
   },
   /**
    * Vue created hook
@@ -118,16 +80,6 @@ export default Vue.extend({
       font-size: 12px;
       letter-spacing: 0.15px;
       text-transform: uppercase;
-    }
-
-    &__since {
-      color: var(--color-text-main);
-      font-weight: bold;
-      font-size: 15px;
-
-      &-days {
-        color: var(--color-text-second);
-      }
     }
   }
 </style>
